# docs/swagger/swagger.yaml
basePath: /api/v1
host: localhost:8080
info:
  contact:
    name: earthcake2233
    url: https://github.com/earthcake2233
  description: cakecake - 轻量级弹幕视频分享平台后端 API
  license:
    name: Custom Non-Commercial License
    url: https://github.com/earthcake2233/cakecake/blob/main/LICENSE
  termsOfService: https://github.com/earthcake2233/cakecake/blob/main/LICENSE
  title: cakecake API
  version: "1.0"
paths:
...
  /space/{userId}/article-favorites:
    get:
      description: Get paginated article favorites for a user space
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: List user article favorites
      tags:
      - Articles
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: JWT Bearer token, e.g. "Bearer {token}"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"